requirements.txt
fastapi
uvicorn
jinja2
python-multipart
pandas
openpyxl
openai
firebase-admin
pypdf
python-docx
reportlab
